+++ CHANGELOG.rst
@@ -3,6 +3,6 @@ Unreleased
 * feat!: cs 69 and flow 3 were removed from offline database.
 * feat: `Introduction.introduction`
 * feat: `Introduction.ombud_messages`
-* feat: `database.boards`
-* feat: `database.cs_codes`
+* feat: `general.boards`
+* feat: `general.cs_codes`
 * fixed: `database.add_instrument_to_db`